# openapi-expanded.yaml
openapi: 3.1.0
info:
  title: OpenNexus Agent API
  description: >
    REST API exposing all AI tools and agents under the OpenNexus Hypercore system.
    Includes core tools, sandboxed environments, AI agents, planners, and MCP dispatcher.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local OpenNexus runtime server
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
      description: "API key to authorize requests"
  schemas:
    CommandRequest:
      type: object
      properties:
        command:
          type: string
      required:
        - command
      example:
        command: "ls -la /workspace"
    OutputResponse:
      type: object
      properties:
        output:
          type: string
      example:
        output: "total 8\n-rw-r--r-- 1 user user 123 Nov 1 09:00 README.md"
    SearchRequest:
      type: object
      properties:
        query:
          type: string
      required:
        - query
      example:
        query: "OpenNexus Hypercore documentation"
    SearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              link:
                type: string
              snippet:
                type: string
      example:
        results:
          - title: "OpenNexus Home"
            link: "https://opennexus.example/docs"
            snippet: "OpenNexus is a hypercore-based runtime..."
    CodeRequest:
      type: object
      properties:
        code:
          type: string
      required:
        - code
      example:
        code: "print('hello world')"
...
  responses:
    BadRequest:
      description: "Invalid request"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: "Missing required field 'command'"
    Unauthorized:
      description: "Unauthorized"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: "API key missing or invalid"
paths:
...
  /tools/bash:
    post:
      operationId: runBash
      summary: Run a bash command
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandRequest"
            examples:
              run:
                summary: Simple ls
                value:
                  command: "ls -la"
      responses:
        '200':
          description: Command output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutputResponse"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
  /tools/web_search:
    post:
      operationId: webSearch
      summary: Perform a web search using DuckDuckGo or Google
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
            examples:
              q:
                summary: Search query example
                value:
                  query: "OpenNexus Hypercore"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        '400':
          $ref: "#/components/responses/BadRequest"
  /tools/python:
    post:
      operationId: runPythonCode
      summary: Execute Python code and return output
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRequest"
            examples:
              hello:
                summary: Hello world example
                value:
                  code: "print('hello world')"
      responses:
        '200':
          description: Output from Python interpreter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutputResponse"
        '400':
          $ref: "#/components/responses/BadRequest"
...